// servidor-local/src/utils/types.d.ts
export interface PrestadorDBType {
    id: string;
    taxa_urgencia: number;
    percentagem_desconto: number;
    minimo_desconto: number;
    nif: string;
    profissao: string;
    enable: boolean;
    created_at: string;
    updated_at: string;
}